libc/string: prevent libc in the kernel/userspace optionally
VELAPLATFO-14058

Add the `LIBC_PREVENT_STRING_KERNEL` and `LIBC_PREVENT_STRING_USER`
that are meant to be selected by the chip if no libc implementation
is going to be built. If selected, neither NuttX's software version
of the libc nor any architecture-specific implementation will be
built in the kernel or in the userspace, respectively. In this
case, the linker may provide a ROM-defined version of the libc
functions instead.

Change-Id: If0ebc66db63845baa064e2a77b6aced7c4d9d384
Signed-off-by: yangdongdong <yangdongdong@xiaomi.com>