arch/xtensa/esp32_esp32s3: prevent arch's libc in the userspace
VELAPLATFO-15382

ESP32 and ESP32-S3 should use the ROM-defined versions of the libc
in flat build and, when building the protected mode, in the kernel.

The ROM-defined version of the libc functions can't be used in the
userspace, however, because it isn't allowed to access the memory
region in flash directly from the userspace. That being said,
`LIBC_PREVENT_STRING_KERNEL` should be selected to avoid building
any implementation of the libc, being the ROM-defined versions
linked instead.

NuttX's software implemented version of the libc will be built in
the userspace. Also, the assembly-defined version of some of the
libc functions (`XTENSA_xxx`) may also be selected to be used in
the userspace.

Change-Id: I55fb73e61e2dc0a0744a853b26981105394d74db
Signed-off-by: yangdongdong <yangdongdong@xiaomi.com>