iob_alloc: change sem_post to count++
if there are two throttled wait, when iob_free occurs, one of wait
will be awakened to execute iob_alloc_committed, but it will fail
to execute, sem will be posted at this time, then another wait will
be awakened. after the other wait thread is awakened, This step is
repeated. the two threads are in the critical_section state and
cannot be switched to other threads. then cpu will busy util timeout.

Change-Id: Ic79fc59ff109e8df2bd214d5a0649d1d5a0572d5
Signed-off-by: zhanghongyu <zhanghongyu@xiaomi.com>