driver/sensor: Solve the busy loop problem caused by sampling problems
WENTAI-8824

Solve the problem that when a user modifies the sampling rate,
the difference between the generation and the mainline generation
may be greater than the own sampling interval.

After this problem occurs, sensor_is_update() will continue to
return true, resulting in a busy loop in the upper-layer application,
so, need update user generation according to mainline generation.

Signed-off-by: dongjiuzhu1 <dongjiuzhu1@xiaomi.com>
Change-Id: Ieb0dc85cd7a0a73ed1685d17eabca2f159030f79