kmm_map.c: Add way to test if addr is within kmap area or not
VELAPLATFO-18841

is_kmap_vaddr is added and used to test that a given (v)addr is actually
inside the kernel map area. This gives a speed optimization for kmm_unmap,
as it is no longer necessary to take the mm_map_lock to check if such a
mapping exists; obviously if the address is not within the kmap area, it
won't be in the list either.

Change-Id: I5d80652ec9f6f7057c1539a75629e3dd8c85e1d0
(cherry picked from commit 5722b8d945b511d7b74e78207f58457aa93b9aa0)