sched/task: Remove spawn_proxyattrs as obsolete implementation
VELAPLATFO-32419

Like the name implies, it is supposed to set the spawn attributes for
the NuttX specific "spawn proxy task" which was historically used as
a proxy to spawn new tasks. The proxy handled file actions and the signal
mask which are inherited from the parent.

The proxy task does not exist anymore, thus the proxy task attributes
do not need to be set anymore either.

Also, the function is currently still used, but the signal mask is set
for the spawning process, not the proxy process, and this is most
DEFINITELY an error (as the spawning process's signal mask changes
unexpectedly).

Setting the signal mask for the newly spawned process is simple, just
set it directly, if instructed to do so. This will be done in a later
patch!

Change-Id: I903439edef66b74bad007644da3281fa8a1dce5b
Signed-off-by: Shoukui Zhang <zhangshoukui@xiaomi.com>