drivers/usbdev: push cdcacm rx buffer to serial rx buffer
VELAPLATFO-22304

When the USB is unplugged during use of cdcacm and
the read req has not been completely processed by
the application, nrdq is not equal to 0.

At this time, the remaining req data needs to be pushed
to the serial rx buffer.

Signed-off-by: dongjiuzhu1 <dongjiuzhu1@xiaomi.com>
Change-Id: I3648ec7acdab2a56cebb626a62b7e6144dd3170f
(cherry picked from commit 7b8f477dd5392a8bdd36497b5158e0148419e3d9)