xtensa_cache:Disable clean/flush optimization in case of SMP restriction
VELAPLATFO-27253

In a multicore task scenario, there may be a situation where the task runs on different cores at different time slices (when the task is not bound to a particular core).
When the task calls cache clean/flush(range > cache size), depending on the optimization, clean_all, flush_all are called. however, at this point, there may be dirty data or incomplete data profiles in the cache on the kernel that is running the task, which may result in dirty data being flushed into memory or make the application think that the flushed data should be successfully flushed into memory, leading to unknown consequences.

Change-Id: Ib5fae1092a58d99aea0050ae570ed29d9c87a077
Signed-off-by: chenrun1 <chenrun1@xiaomi.com>
(cherry picked from commit 75466cdee2ea67192991b11ba364f7ecf541986b)