signal: fix deadlock when sigdeliver call enter_critical_section
VELAPLATFO-28133

cpu0                                 cpu1:

user_main
signest_test
sched_unlock
nxsched_merge_pending
nxsched_add_readytorun
up_cpu_pause
			             arm_sigdeliver
				     enter_critical_section

Reason:
In the SMP, cpu0 is already in the critical section and waiting for cpu1 to enter the suspended state.
However, when cpu1 executes arm_sigdeliver, it is in the irq-disabled state but not in the critical section.
At this point, cpu1 is unable to respond to interrupts and
is continuously attempting to enter the critical section, resulting in a deadlock.

Resolve:
adjust the logic, do not entering the critical section when interrupt-disabled.

Change-Id: I7e15689b2c4081ada70d3bd4c6436c509f698e7b
Signed-off-by: hujun5 <hujun5@xiaomi.com>
(cherry picked from commit 1c129c90bbfc4cd999eae4daec51bfc014da3283)