ipfrag: remove assert for actual product reasons
VELAPLATFO-29715

If the device's CPU resources are scarce and unable to execute the
ip_fragin_timerwork callback in a timely manner, this assert will
be triggered. This is a normal scenario that can occur. The logic
should be modified to wait for the ip_fragin_timerwork to be
executed if it has not been executed yet.

Change-Id: I57f24128c08bc14bca7ab1164ea197a7729aa2ed
Signed-off-by: zhanghongyu <zhanghongyu@xiaomi.com>