drivers/sensors: avoid using sre lock between rptun thread and hpwrok
O62WT-7153

When the system cannot obtain buffers for a long time, hpwork will hold
a lot of buffers. However, due to a deadlock between the rptun thread
and hpwork, hpwork cannot send the buffers in time, causing the system
to crash.

so, avoid hold sre lock in rptun thread.

Change-Id: Idfb6e0ab035625b7d80b2e5658363079a7c3ea37
Signed-off-by: dongjiuzhu1 <dongjiuzhu1@xiaomi.com>