elf:just use phdrs load elf
VELAPLATFO-46833

If a compressed file system is used, the seek back performance of
some compression algorithms is very poor.
Because the section header is at the end of the elf file,
the performance is poor. So we only use the program header for parsing.
And we will sort according to the program header and then load it,
trying not to do the seek back operation on the file.

Change-Id: I86b9e54a39d723addb95e3ee1669f73ec60ce8cf
Signed-off-by: anjiahao <anjiahao@xiaomi.com>